rename runtime programs to clients
